fix: fixed wrong agent list layout
